requirements.txt
requests
discord.py
asyncio
